modify comments and add regression tests
